reworked socket creation/connection, running locally no longer requires internet connection

+++ scripts/mb.pl
@@ -231,10 +231,15 @@ foreach my $gene (@genes) {
 print "done.\n\n";
 
 # Returns the external IP address of this computer
-chomp(my $server_ip = `dig +short myip.opendns.com \@resolver1.opendns.com`);
+chomp(my $server_ip = `dig +short myip.opendns.com \@resolver1.opendns.com 2>&1`);
+if ($server_ip !~ /(?:[0-9]{1,3}\.){3}[0-9]{1,3}/) {
+	print "Could not determine external IP address, only local clients will be created.\n";
+	$server_ip = "127.0.0.1";
+}
+
 # Initialize a server
 my $sock = IO::Socket::INET->new(
-	LocalAddr  => $server_ip.":".$port,
+	LocalPort  => $port,
 	Blocking   => 0,
 	Reuse      => 1,
 	Listen     => SOMAXCONN,
@@ -249,7 +254,13 @@ print "Job server successfully created.\n";
 chomp(my $server_hostname = `hostname`);
 if (scalar(@machines) == 0) {
 	push(@machines, $server_hostname);
-	$machines{$server_hostname} = $server_ip;
+	$machines{$server_hostname} = "127.0.0.1";
+}
+elsif (scalar(@machines) == 1) {
+	# Check if the user input only the local machine in the config
+	if ($machines{$machines[0]} eq $server_ip) {
+		$machines{$machines[0]} = "127.0.0.1";
+	}
 }
 
 my @pids;
@@ -265,7 +276,7 @@ foreach my $machine (@machines) {
 		(my $script_name = $script_path) =~ s/.*\///;
 
 		# Move required datafiles to machines, initialize clients
-		if ($machines{$machine} ne $server_ip) {
+		if ($machines{$machine} ne "127.0.0.1" && $machines{$machine} ne $server_ip) {
 			# Send this script to the machine
 			system("scp", "-q", $script_path, $machine.":/tmp");
 
@@ -284,7 +295,7 @@ foreach my $machine (@machines) {
 			system("cp", $mb, "/tmp");
 
 			# Execute this perl script on the given machine
-			exec("perl", "/tmp/$script_name", "--server-ip=$server_ip");
+			exec("perl", "/tmp/$script_name", "--server-ip=127.0.0.1");
 		}
 
 		exit(0);
@@ -473,7 +484,11 @@ sub client {
 
 	# Determine this host's IP
 	chomp(my $ip = `dig +short myip.opendns.com \@resolver1.opendns.com`); 
-	die "Could not establish an IP address for host.\n" if (not defined $ip);
+
+	# Set IP to localhost if we don't have internet
+	if ($ip !~ /(?:[0-9]{1,3}\.){3}[0-9]{1,3}/) {
+		$ip = "127.0.0.1";
+	}
 
 	# Spawn more clients
 	my @pids;
@@ -541,7 +556,7 @@ sub client {
 			unlink(@results);
 
 			# Send the results back to the server if this is a remote client
-			if ($server_ip ne $ip) {
+			if ($server_ip ne "127.0.0.1" && $server_ip ne $ip) {
 				send_file({'FILE_PATH' => $gene_archive_name, 'FILE_HANDLE' => $sock});	
 				unlink($gene_archive_name);
 			}
